requirements.txt
ultralytics==8.3.0
torch
torchvision
numpy
opencv-python
pillow
matplotlib
streamlit
